refactor: rename utility file
